+++ packages/sdk-codegen/src/typescript.gen.ts
@@ -81,10 +81,10 @@ import { sdkVersion } from '../constants'
 import { IDictionary, ${this.typeNames().join(', ')} } from './models'
 
 export class ${this.packageName} extends APIMethods {
-
+  static readonly ApiVersion = '${this.apiVersion}'
   constructor(authSession: IAuthSession) {
     super(authSession, sdkVersion)
-    this.apiVersion = '${this.apiVersion}'
+    this.apiVersion = ${this.packageName}.ApiVersion
     this.apiPath =
       authSession.settings.base_url === ''
         ? ''
@@ -106,9 +106,10 @@ import { sdkVersion } from '../constants'
 import { IDictionary, ${this.typeNames(false).join(', ')} } from './models'
 
 export class ${this.packageName}Stream extends APIMethods {
+  static readonly ApiVersion = '${this.apiVersion}'
   constructor(authSession: IAuthSession) {
     super(authSession, sdkVersion)
-    this.apiVersion = '${this.apiVersion}'
+    this.apiVersion = ${this.packageName}Stream.ApiVersion
     this.apiPath =
       authSession.settings.base_url === ''
         ? ''