+++ packages/sdk/src/3.1/streams.ts
@@ -241,9 +241,10 @@ import {
 } from './models'
 
 export class Looker31SDKStream extends APIMethods {
+  static readonly ApiVersion = '3.1'
   constructor(authSession: IAuthSession) {
     super(authSession, sdkVersion)
-    this.apiVersion = '3.1'
+    this.apiVersion = Looker31SDKStream.ApiVersion
     this.apiPath =
       authSession.settings.base_url === ''
         ? ''